taxon-tree: expand pinned family tree by default
